refactor(query): extract Canary.Query.Health read model
Move from lib/canary/query.ex into lib/canary/query/health.ex:
- Public: health_targets/0, health_status/0, target_checks/2,
  recent_transitions/1
- Exposed as @doc false public for facade reuse:
  recent_transitions_since/1
- Privates: fetch_recent_checks/2
- Module attr: @recent_checks_limit

Facade exposes health functions via defdelegate and calls
Canary.Query.Health.recent_transitions_since/1 in report_slice/1.
Target, TargetCheck, TargetState aliases dropped from the facade.
