+++ lib/canary/query.ex
@@ -6,10 +6,7 @@ defmodule Canary.Query do
 
   alias Canary.Schemas.{
     Error,
-    ErrorGroup,
-    Target,
-    TargetCheck,
-    TargetState
+    ErrorGroup
   }
 
   import Ecto.Query
@@ -180,58 +177,10 @@ defmodule Canary.Query do
     }
   end
 
-  @recent_checks_limit 5
-
-  @spec health_targets() :: [map()]
-  def health_targets do
-    repo = Canary.Repos.read_repo()
-
-    targets_with_state =
-      from(t in Target,
-        left_join: s in TargetState,
-        on: t.id == s.target_id,
-        order_by: t.name,
-        select: {t, s}
-      )
-      |> repo.all()
-
-    target_ids = Enum.map(targets_with_state, fn {t, _} -> t.id end)
-
-    checks_by_target = fetch_recent_checks(repo, target_ids)
-
-    Enum.map(targets_with_state, fn {target, state} ->
-      recent = Map.get(checks_by_target, target.id, [])
-
-      %{
-        id: target.id,
-        name: target.name,
-        service: Target.service_name(target),
-        url: target.url,
-        state: (state && state.state) || "unknown",
-        consecutive_failures: (state && state.consecutive_failures) || 0,
-        last_checked_at: state && state.last_checked_at,
-        last_success_at: state && state.last_success_at,
-        latency_ms: recent |> List.first() |> then(&(&1 && &1.latency_ms)),
-        tls_expires_at: Enum.find_value(recent, & &1.tls_expires_at),
-        recent_checks:
-          Enum.map(recent, fn c ->
-            %{
-              checked_at: c.checked_at,
-              result: c.result,
-              status_code: c.status_code,
-              latency_ms: c.latency_ms
-            }
-          end)
-      }
-    end)
-  end
-
-  @spec health_status() :: map()
-  def health_status do
-    targets = health_targets()
-    summary = Canary.Summary.health_status(%{targets: targets})
-    %{summary: summary, targets: targets}
-  end
+  defdelegate health_targets(), to: Canary.Query.Health
+  defdelegate health_status(), to: Canary.Query.Health
+  defdelegate target_checks(target_id, window), to: Canary.Query.Health
+  defdelegate recent_transitions(window), to: Canary.Query.Health
 
   @spec error_groups(String.t()) :: {:ok, [map()]} | {:error, :invalid_window}
   def error_groups(window) do
@@ -247,13 +196,6 @@ defmodule Canary.Query do
     end
   end
 
-  @spec recent_transitions(String.t()) :: {:ok, [map()]} | {:error, :invalid_window}
-  def recent_transitions(window) do
-    with {:ok, cutoff} <- Canary.Query.Window.to_cutoff(window) do
-      {:ok, recent_transitions_since(cutoff)}
-    end
-  end
-
   @spec report_slice(String.t()) :: {:ok, map()} | {:error, :invalid_window}
   def report_slice(window) do
     with {:ok, cutoff} <- Canary.Query.Window.to_cutoff(window) do
@@ -262,52 +204,8 @@ defmodule Canary.Query do
          error_groups: error_groups_since(cutoff),
          error_summary: error_summary_since(cutoff),
          incidents: active_incidents(),
-         recent_transitions: recent_transitions_since(cutoff)
+         recent_transitions: Canary.Query.Health.recent_transitions_since(cutoff)
        }}
-    end
-  end
-
-  # Batch-fetch top-N recent checks per target using ROW_NUMBER window function.
-  # 1 query replaces N individual queries.
-  defp fetch_recent_checks(_repo, []), do: %{}
-
-  defp fetch_recent_checks(repo, target_ids) do
-    ranked =
-      from(c in TargetCheck,
-        where: c.target_id in ^target_ids,
-        select: %{
-          target_id: c.target_id,
-          checked_at: c.checked_at,
-          result: c.result,
-          status_code: c.status_code,
-          latency_ms: c.latency_ms,
-          tls_expires_at: c.tls_expires_at,
-          rn: over(row_number(), :w)
-        },
-        windows: [w: [partition_by: c.target_id, order_by: [desc: c.checked_at]]]
-      )
-
-    from(r in subquery(ranked),
-      where: r.rn <= ^@recent_checks_limit,
-      order_by: [asc: r.target_id, asc: r.rn]
-    )
-    |> repo.all()
-    |> Enum.group_by(& &1.target_id)
-  end
-
-  @spec target_checks(String.t(), String.t()) ::
-          {:ok, [%TargetCheck{}]} | {:error, :invalid_window}
-  def target_checks(target_id, window) do
-    with {:ok, cutoff} <- Canary.Query.Window.to_cutoff(window) do
-      checks =
-        from(c in TargetCheck,
-          where: c.target_id == ^target_id and c.checked_at >= ^cutoff,
-          order_by: [desc: c.checked_at],
-          limit: 500
-        )
-        |> Canary.Repos.read_repo().all()
-
-      {:ok, checks}
     end
   end
 
@@ -428,22 +326,6 @@ defmodule Canary.Query do
         unique_classes: count(g.group_hash)
       },
       order_by: [desc: sum(g.total_count)]
-    )
-    |> Canary.Repos.read_repo().all()
-  end
-
-  defp recent_transitions_since(cutoff) do
-    from(t in Target,
-      join: s in TargetState,
-      on: t.id == s.target_id,
-      where: s.last_transition_at >= ^cutoff,
-      order_by: [desc: s.last_transition_at, asc: t.name],
-      select: %{
-        target_id: t.id,
-        target_name: t.name,
-        state: s.state,
-        transitioned_at: s.last_transition_at
-      }
     )
     |> Canary.Repos.read_repo().all()
   end
